pfx: expose enough of pfxt to make it independently usable
