+++ rtrlib/rtr/packets.c
@@ -17,7 +17,7 @@
 #include "rtrlib/lib/convert_byte_order_private.h"
 #include "rtrlib/lib/log_private.h"
 #include "rtrlib/lib/utils_private.h"
-#include "rtrlib/pfx/trie/trie-pfx_private.h"
+#include "rtrlib/pfx/pfx_private.h"
 #include "rtrlib/rtr/packets_private.h"
 #include "rtrlib/rtr/rtr_private.h"
 #include "rtrlib/spki/hashtable/ht-spkitable_private.h"
